Solve Part 2: Crack salted Passwords
The description for this task was not good.

The function should take an optional second
argument named use_salts. If set to true,
each salt string from the file known-salts.txt
should be appended AND prepended to each password
from top-10000-passwords.txt before hashing and
before comparing it to the hash passed into the
function.

What it actually meant was, take one salt.
Append it to a password, hash it, compare it to
the input hash. If it matches, fine. If not,
take that salt again, this time prepend it, match
it again. Do that for every salt.

Fortunately I found a forum post that made that
clear. Very bad description of what to do.
Implementing that was fairly easy then.